Add gerrit tool for our benefit
Simple to run - just go into one of our directories with a th remote already established
Run this script, it will create a gerrit remote as well as default reviewers and set up the default push to kitkat
So all you have to do for each commit is "git push gerrit" and voila!

Caveat - this is a one-time-only deal for each directory with a git. Things break after the first time you do this.
Z

Change-Id: Ic436d7d184f60675d97b7b8ba45458ef7fae100e