OTA: Add OTA support
A few things are needed to get this working properly. You will need
a flag added to your ~/.bashrc to set the build type. This has
been setup specifically for XenonHD server build paths, so some
modification will be needed for others.

The flags you need is:

export OTA_TYPE=

This tells us what type of build this will be. XenonHD's server uses
build type as part of the download zip path.
At this time, usable values are:

experimental
weeklies
stable

If no value is set, a default of

export OTA_TYPE=experimental

will be used.

XenonOTA also requires two files, ota_conf which is used by the app to read
currently installed build info. The other is ota_lollipop.xml which is
used to read the most current build available for download on the server.
Both of these are automatically generated and no interaction is needed for
ota_conf. ota_lollipop.xml will be added to your out dir. You will
need to move this to a location that will be static. XenonOTA uses this info
to compare what you have vs what is the latest build. For XenonHD we move
this xml to ftp/OTA/<device>, then we move the zips as normal to the
correct location. For example, we move an experimental i605 build to
ftp/experimental/i605. The xml file that is in ftp/OTA/i605 lists build
version and location path of the build download. If this is newer than
what you currently have, it will display an update notification and you
will be able to download the file zip and md5 with the provided dynamic
links.

ota_lollipop.xml has to be hosted at a location where the url doesnt change
and a location that will auto download the xml file vs just opening it to
read. Not all hosts allow this.

Change-Id: Ie0240698156346a9c91fbde5c1b0c194ae2d4784