xenonhd: Introduce (legacy) release-keys ROM signing hack
Turns out that not every single machine is happy about signing created
target files package immediately as some will get out-of-memory error(s)
in RAM as soon as the package is being created.

Introduce a hack that immediately signs all targets with a provided,
"fake" test-keys to workaround this. It's not recommended to enable this
hack unless you're owning one of those unlucky machines.

Change-Id: Ie8b65c49d3a83811c7d1c9342254e7f0fa75d884
Signed-off-by: Albert I <krascgq@outlook.co.id>